+++ src/tensor/tensor_change_dimension.cc
@@ -27,7 +27,7 @@ static const Tensor change_dimension_inner(const Tensor &a, int dim,
                                            index new_size) {
   typedef typename Tensor::elt_t elt_t;
 
-  dim = normalize_index(dim, a.rank());
+  dim = (int)normalize_index(dim, a.rank());
   Indices d = a.dimensions();
   index old_size = d[dim];
   if (old_size == new_size) return a;
